+++ src/api-early.ts
@@ -8,7 +8,7 @@ const { apiToken } = getPreferenceValues() as Preferences;
 export const apiGetCurrentTracking = () =>
   Promise.resolve()
     .then(() => console.debug("api get current tracking"))
-    .then(() => fetch("https://api.timeular.com/api/v3/tracking", params))
+    .then(() => fetch("https://api.early.app/api/v3/tracking", params))
     .then(jsonFromResponse)
     .then(json => (json as { currentTracking: Tracking }).currentTracking);
 
@@ -17,7 +17,7 @@ export const apiStartTracking = ({ activityId, startedAt = new Date() }: { activ
     .then(() => console.debug(`api start tracking activityId ${activityId}`))
     .then(() =>
       fetch(
-        `https://api.timeular.com/api/v3/tracking/${activityId}/start`,
+        `https://api.early.app/api/v3/tracking/${activityId}/start`,
         makeParams("POST", { startedAt: date(startedAt) })
       )
     )
@@ -28,9 +28,7 @@ export const apiStartTracking = ({ activityId, startedAt = new Date() }: { activ
 export const apiStopTracking = ({ stoppedAt = new Date() }: { stoppedAt?: Date }) =>
   Promise.resolve()
     .then(() => console.debug("api stop tracking"))
-    .then(() =>
-      fetch("https://api.timeular.com/api/v3/tracking/stop", makeParams("POST", { stoppedAt: date(stoppedAt) }))
-    )
+    .then(() => fetch("https://api.early.app/api/v3/tracking/stop", makeParams("POST", { stoppedAt: date(stoppedAt) })))
     .then(jsonFromResponse)
     .catch(e => {
       const message = e.message.toLocaleLowerCase();
@@ -44,21 +42,21 @@ export const apiStopTracking = ({ stoppedAt = new Date() }: { stoppedAt?: Date }
 export const apiListAllActivities = () =>
   Promise.resolve()
     .then(() => console.debug("api list all activities"))
-    .then(() => fetch("https://api.timeular.com/api/v3/activities", params))
+    .then(() => fetch("https://api.early.app/api/v3/activities", params))
     .then(jsonFromResponse)
     .then(json => (json as { activities: Activity[] }).activities);
 
 export const apiListAllTagsAndMentions = () =>
   Promise.resolve()
     .then(() => console.debug("api list all tags and mentions"))
-    .then(() => fetch("https://api.timeular.com/api/v3/tags-and-mentions", params))
+    .then(() => fetch("https://api.early.app/api/v3/tags-and-mentions", params))
     .then(jsonFromResponse)
     .then(json => json as { tags: Tag[]; mentions: Mention[] });
 
 export const apiGetTimeTrackingEntries = ({ from, to }: { from: Date; to: Date }) =>
   Promise.resolve()
     .then(() => console.debug(`api get time tracking entries from ${from} to ${to}`))
-    .then(() => fetch(`https://api.timeular.com/api/v3/report/data/${date(from)}/${date(to)}`, params))
+    .then(() => fetch(`https://api.early.app/api/v3/report/data/${date(from)}/${date(to)}`, params))
     .then(jsonFromResponse)
     .then(json => json as { timeEntries: TimeEntry[]; message?: string })
     .then(json => (json.message ? Promise.reject(new Error(json.message)) : json.timeEntries));
@@ -74,7 +72,7 @@ type EditTracking = {
 export const apiEditTracking = (body: EditTracking) =>
   Promise.resolve()
     .then(() => console.debug(`api edit tracking with ${JSON.stringify(body)}`))
-    .then(() => fetch("https://api.timeular.com/api/v3/tracking", makeParams("PATCH", body)))
+    .then(() => fetch("https://api.early.app/api/v3/tracking", makeParams("PATCH", body)))
     .then(jsonFromResponse)
     .then(json => json as { currentTracking: Tracking; message?: string })
     .then(json => (json.message ? Promise.reject(new Error(json.message)) : json.currentTracking));
@@ -86,10 +84,10 @@ type CreateTagParams = {
   spaceId: string;
 };
 
-export const apiCreateTag = ({ spaceId, label, key, scope = "timeular" }: CreateTagParams) =>
+export const apiCreateTag = ({ spaceId, label, key, scope = "early" }: CreateTagParams) =>
   Promise.resolve()
     .then(() => console.debug(`api creating tag #${label}, spaceId ${spaceId}, key ${key || "auto"} in scope ${scope}`))
-    .then(() => fetch("https://api.timeular.com/api/v3/tags", makeParams("POST", { spaceId, label, key, scope })))
+    .then(() => fetch("https://api.early.app/api/v3/tags", makeParams("POST", { spaceId, label, key, scope })))
     .then(jsonFromResponse)
     .then(json => json as Tag);