+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# Timeular Changelog
+# EARLY Changelog
+
+## [Update API endpoint] - 2025-09-16
+
+- The API endpoint has been updated to reflect the product name change from timeular to early.
 
 ## [Update "Edit Note" View] - 2023-02-03
 
@@ -8,6 +12,6 @@
 
 - The "edit" view now supports a textarea, to better capture long form notes in a more friendly way
 
-## [Added Timeular] - 2022-01-23
+## [Added EARLY] - 2022-01-23
 
 Initial version code